Trigger config replace on cert changes
Certs and configs are not related from an API
perspective but are from a deployment perspective
in that a change to a cert needs to trigger a
config change as the deployment needs to start
leveraging the new cert.

This change just handles it for terraform as
someone can use this example code to do the
following:
- Change AKV cert.
- This triggers a change in NGINXaaS certificate.
- With this change, it will auto-trigger a conf
  replacement.

This is a workaround until we have cert rotation
and then we can leverage versionless certs.